move to correct util package
